#117 Evaluate the HTTP status code from scanner response
For better debugging we also evaluate the HTTP status code.

We do the trick by appending it at the end of the response body
with cURL's built in option -w and then extract and remove this
last line.

Why exit code 4?

It looks like the exit codes are incremented for each error condition.
That implies that we should use 4 and increment all following error
codes. But this would not be backward compatible if a caller evaluates
the exit codes. Also we think that the evaluation of the HTTP status
code is the same error condition category like the absent of a process
ID in the response JSON.

This patch closes #117.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>