#50 Add components overview diagram
This diagramm shows a very high level overview of main SCBv2
compononents and how they communicate with each other.

This diagram only shows the most important parts to be as simple
as possible. It mainly uses terms from Kubernetes because the
new architecture is completely based on Kubernetes.

The diagram (uxf) itself is editable with UMLet: https://www.umlet.com/
It isn't made with PlantUML because it was a big hassle to lign up
the boxes properly with PlantUMLs auto layouting.

The rendered image is also commited so one does not need to install
UMLet first to view it.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>