Move adr of securecodebox.io to main repo
Since we changed the framework for our website the decision described in
the respective adr file will be archived or forgotten about. Furthermore
this website is a tool we use for our main project, thus the decision
about what "tool" we use for documentation should be saved in the main
repository.