Start minikube in Vagrant setup
The Vagrant user requires the docker group to use the Docker provider.
Due to the fact that this group is created in the same login session
the groop is not available w/o a re-login. So as workaround sudo is
used.

The STDERR is redirected to /dev/null because Minikube prints out download
progess bars and this will clutter the Vagrant output massively.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>