Ensure that crds are only contained once in the helm2 chart
Copying them over makes them appear twice in the chart, which doesn't work if helm3 is used to insall the chart build for helm2.
This is sometimes required for organisations moving from helm 2 to 3.

This should not affect the helm3 build as the helm2 build happens after the helm3 build.