Del leading 'v' In git-tag To create valid Semver
We noticed that our current release process did not use the right docker
images anymore. This problem was created after switching to the new
build and push docker action because the images are now tagged with a
valid semver instead of a version with a leading 'v' but the release
tried to use the 'old' version style with a leading 'v'.

To resolve this issue we decided to trim the leading 'v' in our Release
Action because we want to have consistent tags in Github

Signed-off-by: Yannik Fuhrmeister <yannik.fuhrmeister@iteratec.com>