Change Templating Engine To Nunjucks
Experimenting with different Templates we encountered the problem that
we had to add a lot of helper methods due to mustache beeing a logic
less templating engine.

This would have meant that it would be necessary to provide new helper
methods for new templates that want to display different information of
scan or findings.

To resolve this issue we decided to switch to the Nunjucks templating
engine

This Commit:
- Switches template engine
- removes now unnecessary helper methods
- write the slack-messageCard template with nunjucks

See: https://mozilla.github.io/nunjucks/

Signed-off-by: Yannik Fuhrmeister <yannik.fuhrmeister@iteratec.com>