Converting the ADRs to markdown
We want to use the ADRs in our documentation project, but this
does not support Asciidoc. So we decided to convert them.

Initial convertion is done w/ Pandoc:

```
for i in {0..7}; do \
  asciidoc -b docbook "adr_000${i}.adoc" \
  pandoc -f docbook -t markdown_strict "adr_000${i}.xml" -o "adr_000${i}.md" \
  rm "adr_000${i}.xml" \
done
```

There were some syntax errors afterwards which were fixed manually.

The linebreaks of paragraphs were removed consequntly because we use
linewrap in the editors.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>