Use Multi-Stage-Build For Nikto
As described in #610 the build for nikto was quite complicated and could
not be managed via matrix builds on GH.

To Fix this I created a multi-stage Dockerfile that clones the nikto
repo in the build stage and checks out the specified version.

The next stage installs all dependencies like in the Dockerfile provided
by the maintainer of nikto, copies nikto from the build stage and our
wrapper script.

I also noticed that the alpine version was a little behind so I changed
it from 3.12 to 3.14

Signed-off-by: Yannik Fuhrmeister <yannik.fuhrmeister@iteratec.com>