Add Pipeline For Release Image Builds
Our goal is to simplify the colaboration with forks.
To do so we want to ensure that the CI pipeline can be executed on a
fork easily. Pushing to an image registry requires that the owner of the
fork provides some secrets to this image registry.

To avoid this problem we decided to split the CI pipeline into a release
pipeline and a developement pipeline. The release pipeline is only
triggered on the release event and is only responsible for building and
pushing the container images. The developement pipeline handles the use
of container images internally and does not rely on any image registry.

This commits adds the initial build pipeline.

NOTICE: The event that triggeres the pipeline is still set to 'push' to
ensure that the pipeline is building the images correctly. This has to
be changed before merging into main

Signed-off-by: Yannik Fuhrmeister <yannik.fuhrmeister@iteratec.com>