Braces around the links causes docusaurus build fail
The ADRs from this repo are copied during the documentation build
over to that repo and will be rendered in its ADR section.

Docusaurus fails the build if there are unresolvable links in
the sources.

In Markdown typically links are written as `[text](url)`. If the
links are written in the footer of the document with references
the syntax must be:

```Markdown
[text]: uri
```

The braces around the URI resulted into unresolvable URIs which
caused the failing build.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>