Simplify the makefile by removing shell specific syntax
It is not necessar to add ;\ at the end of each line because make
runs each command step by step. So this is removed.

Also echo what is done is not necessary because make prints by
default each command for debug purpose.

Also removed the cd <subdir> because this is brittle: If the cd
commandfails the subsequent commands work in a wrong directory.
Better solution is to always use theconcrete path relative to
the Makefile.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>