Remove non-POSIX compliant switch from sed
The -i Switch (in place file processing) is not POSIX compliant
and behaves differently on some systems:

* BSD Unix: requires an argument (e.g. -i '.bak')
* GNU Linux: does not allow an argument

This patch introduces a platform independent way for replacing
the needle stringin with the scanner name in the copied template
files.

The trick is to use a temporary file to save the sed result. First
aproach was to use > to redirect the sed output from STDOUT into
the temp file. This does not work because find's -exec does not
execute a shell and executes the command directly in a forked
child process. So there are no shell features like redirect or
pipe. The solution is to use the sed command 'w' to write the file
and supress the output on STDOUT.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>