Add templating for envs in cascading scans
Cascading scan definitions use templating to include dynamic information
in the scans they trigger. This templating was previously not applied
to the definitions of environmental variables. However, sometimes it may
be desireable to have templated env variables, as certain information
can be easier to pass to the system as an env var, as opposed to a
parameter.

Signed-off-by: Max Maass <max.maass@iteratec.com>