Allow reposcanner to output latest commit id
In some cases, the latest commit ID (SHA) of the found repositories can
be an interesting piece of metadata (for example, to annotate the
version field for DefectDojo output in cascading rules).
This commit adds this functionality to git-repo-scanner. It is disabled
by default, as it causes an extra API hit per detected repository.

Signed-off-by: Max Maass <max.maass@iteratec.com>