Respect affinity and tolerations from helm values
This commit makes the operator respect affinity and tolerations that are
specified in ScanSpec and ParseDefinitions. The default values are used
if the scan does override it, and otherwise discarded (not merged).
Default values are not inherited like explicitly specified values are,
and you need to make sure to specify affinity and tolerations for both
scanner and parser (they are not copied over, either).
Hooks will get the same treatment in the next commit.

Signed-off-by: Max Maass <max.maass@iteratec.com>