Let parser construct repo URLs from annotation
Before Gitleaks 8.0, the parser used to construct a direct URL to
each detected commit based on the parameter used to clone the repo.
Since it can no longer clone repos, this is no longer feasible. However,
this commit adds the capability to pull the repo information from a scan
annotation and use that. It does not actually enforce that the provided
repository URL matches the one that was cloned in the init container -
it blindly trusts whatever data it is given.

Signed-off-by: Max Maass <max.maass@iteratec.com>