#754 Extract thew hole generation code into a shell script
Because it is shell script anyway and editing shell script in
Makefiles is tedious I decided to extract the code into a separate
shell script. So we can invoke this from the make targets.

AFAICS there is lots of duplicatedcode for hooks and parser. Seems
to be nearly the same as for the scanners. So we may reuse thisscript.

This will elad into a cleaner Makefile and less code to maintain.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>