Embed original DefectDojo finding in synced back findings
See #1055

This might be extended / improved in future by reducing the number of api requests performed by this feature. At the moment the hook needs to do 1 api request for every duplicate findings. The DefectDojo API would already support to prefetch duplicate findings in the list api request but this is not yet supported by our DefectDojo Java API client. Also might be pretty tricky to do with the current way the client is written.

Co-authored-by: Max Maass <max.maass@iteratec.com>
Co-authored-by: Sebastian Franz <sebastian.franz@iteratec.com>
Signed-off-by: Jannik Hollenbach <jannik.hollenbach@iteratec.com>