Add support for *-plus ZAP report formats
ZAP reports can come in an enhanced "plus" format, which also contains
the full request and response headers and payload.  This commit adds
support for selecting these formats to ZAP Advanced.

To enable them, simply add '-r XML-plus' to the parameters in the Scan
definition.  The XML-plus format is compatible with the regular XML
format used by the parser.

Signed-off-by: Max Maass <max.maass@iteratec.com>