#1353 Extract npm-ci-all shell code in script
Since make can't deal with spaces in pathsm it is not feasable to
find the packages and iterate over them in make. I tried to set
IFS in make: Does neither work with for nor while.

As a workaround the whole shell code for this target is moved into
a shel script which can deal with spaces in paths.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>