#1353 Document and fix indentation problem of conditionals in make
Problem is, if the body of a conditional is indented like:

```
ifeq (1,1)
	$(error SNAFU!)
endif
```

and the indentation is done with tabs (spaces didn't produce the eror)
then the script fails on macOS/FreeBSD because the indented line is
interpreted as command. GNU Make on macOS/FreBSD is more strict about
that than GNU Make on GNU/Linux system. The script fails with:

> Makefile:8: *** commands commence before first target.  Stop.

This means that a command (with tab indented line) must be written
after a recipe (target).

The solution is to remove all indentation in conditional blocks.
The solution to use spaces was discarded because it is errorprone
to accidentally convert them to tabs and introduce a regression,
even if the readability of such conditionals suffer.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>