#911 Create a download cache for trivy database
This uses the trivy client/server functionality to download the
vulnerability DB only once for multiple scans to avoid GitHub's rate
limit. An additional container running trivy in server mode is started,
and all scans connect to it. This integrates the functionality
previously documented [in the SCB docs for trivy][1], now that trivy
supports remote scanning for more than just container image scans.

[1]: https://www.securecodebox.io/docs/scanners/trivy/#scanning-many-targets

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>