#1489 Fixes the tests of Trivy and adds a check for empty scanResults
Trivy tests have been ruuning incorrectly until now. The JSON.parse function was missing, that means the tests for empty results  was succesfful because
of the earlier check of "if the results are a string or not" and not the check of "results is an non-empty array"
This fixes that.

Signed-off-by: Ilyes Ben Dlala <ilyes.bendlala@iteratec.com>