#1838 Use trivy to generate CycloneDX SBOMs
Add a new ScanType to trivy, that allows generating CycloneDX SBOMs from
container images. This is part of work to integrate an SBOM workflow
into the secureCodeBox. At the moment the SBOMs are only generated and
uploaded to storage, nothing else happens to them.

Note that this is a different result type than the other trivy scans,
this uses "sbom-cyclonedx", therefore the normal trivy parser will not
run for these scans.

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>