#1838 Add a parser for CycloneDX SBOMs
This is a bit tricky, because now trivy can generate two different kinds
of scan results, normal trivy-json results and sbom-cyclonedx results.
To keep these parsers separated, an additional CycloneDX parser is
added. To not overcomplicate the structure, this parser is added in the
trivy scanner directory. If in the future there are multiple scanners
creating SBOMs, it can be moved somewhere else.

Adding a second parser for a scanner requires some Makefile trickery.
Usually the docker-build, docker-export, kind-import and deploy Makefile
targets are used to build and deploy a scanner. Adapting the first three
targets is easy, the additional parser just needs an extra build step
wich can be added as another dependency. The deploy is more tricky
because it is a single helm upgrade --install command, which sets the
necessary values for docker repository and tag. This is achieved by
redefining the Makefile target, but it is not the nicest solution, make
warns about the target getting overridden.

The parser is a generic cyclonedx parser (although with currently
limited functionality), but currently lives in the trivy scanner
directory because it is only used here and means the global structure
does not have to get changed. Because it is not trivy-specific, it is
not supposed to use "trivy" in its name/repository. This means, the
common-... Makefile targets cannot always be used, and if they can, they
have to be used in unintended ways. By default they get used by setting
the module variable to the directory name of what is built and then add
the name of the scanner in the end, for example the parser for trivy
ends up being built as parser-trivy from the parser directory. Now to
prevent the CycloneDX parser getting named parser-cyclonedx-trivy, the
docker-build-cyclonedx-parser target cannot use the common-docker-build
and the export/import targets have to be a bit creative with using the
module and name variables. This might get reworked to make the Makefiles
nicer again.

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>