#1838 Add a Dependency-Track persistence hook
Basic functionality of a persistence hook, that sends CycloneDX SBOMs to
OWASP Dependency-Track. Since hooks run for all scans, the hook first
checks the scanType and then the bomFormat property to see if the
rawResult is actually a CycloneDX SBOM. If it is it gets uploaded to
Dependency-Track.

To assign an SBOM to a project, either the project UUID or the name and
version are required. The hook tries to grab the name and version from
the name property of the SBOM and automatically creates missing
projects. This is not the nicest way to get name and version, but there
is no better property in the SBOM to read it from. It works for
container scans, but probably breaks for other kinds of SBOM scans.

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>