#1838 Add unit tests for dependencytrack hook
Make sure the persistence-dependencytrack hook can be tested
automatically by providing some tests for it. Some parts of the tests
might be coupled a bit tightly to the actual implementation, if the
endpoint is switched to the PUT endpoint the tests will have to be
adapted.

The tests inject the mocked dependencies through parameters, similar to
how the tests for the generic webhook work. Since fetch() is the only
dependency here, this works pretty well.

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>