#1838 Add unit tests for parser-cyclonedx
Add a small and very simple test for the CycloneDX parser, after all it
doesn't do much. More annoying is again the Makefile structure, since
the project is not set up for multiple parsers for a scanner, it also
does not accommodate that for the tests. To circumvent this, the added
target for the parser-cyclonedx tests needs to manually run the
install-deps-js target with a different module name, so that all the
dependencies get installed correctly.

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>