#1838 Fix image name and version recognition
The previous way of detecting the name and version to supply to
Dependency-Track was very brittle, it already failed for image
references including a hash, resulting in names like hello-world@sha256,
because it would only split on ':' and then select the first and last
component.

This version uses a regex to as accurately as possible match the
individual components of a docker image reference. The regex comes from
the [official implementation] on GitHub, but is actually taken from a
[pull request], which adds named capture groups and fixes an issue with
domain recognition being too eager.

Yes the regex looks pretty wild, yes there are tests. I don't think it
makes sense to build the regex from the individual components like the
docker library does it.

Unfortunately this does not solve the problem of actually getting the
reference from somewhere, for images it works with getting it from the
name of the main component, but this part stays brittle. Annotations to
the scans might be a possible solution for that.

[official implementation]: https://github.com/distribution/reference
[pull request]: https://github.com/distribution/distribution/pull/3803

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>