#1894 React to container stop events as well
Previously the AWS monitor would only react to RUNNING state changes,
now it also detects STOPPED events and also forwards those to the
kubernetes Reconciler. This is still at a Task-level, which works, but
is somewhat generic, there might be ScheduledScans that would not be
needed anymore if the containers have already stopped.

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>