#1894 Count container instances per image
Remember how many containers using one image are running and only delete
the scan when the last one has been deleted. This prevents the scan from
getting deleted after the first instance of an image stops and instead
defers it till the last one stops.

Kind of moot at this point because this MVP still uses scans instead of
scheduled scans, but needed for heading in the right direction.

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>