#1894 Create ScheduledScans instead of Scans
The point of the AutoDiscovery is to continuously monitor the resources
it tracks. This is done by creating ScheduledScans when they appear and
deleting them when they disappear.

For SBOMs this is not particularly useful, in most cases (excluding
upgrades to trivy-sbom), generating an SBOM for the same image and
digest should result in the same SBOM. Generating only Scans for SBOM
targets instead of ScheduledScans would make everything a lot more
complicated though, so they are also covered by ScheduledScans.

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>