#1894 Add basic integration tests for kubernetes
Use envtest to test the kubernetes part of the service against an actual
kubernetes control plane. This is the same way the kubernetes
autodiscovery and the operator are tested, but here it is used without
controller-gen or kustomize because this is not a kubebuilder project
without any custom resources.

The current tests are only a small demonstration to connect to the
envtest control plane. It is not clear yet where the envtest connection
should be made, in the tests for the kubernetes module, or the tests for
main. To me it seems like the tests for the kubernetes module should
test only that, but the tests for main should test the combination of
the aws and kubernetes parts, and then both test suites need to access a
kubernetes control plane. If envtest supports multiple different ones
that could work, otherwise it might need to be shared.

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>