#1894 Make scans and other options configurable
Previously the ScheduledScans the autodiscovery would create were
hardcoded to trivy-sbom-image. Even for an MVP this is not that helpful
though. This makes the scans and some other options configurable in much
the same way they are for the kubernetes autodiscovery, by reading from
a yaml config file which gets supplied to a commandline flag.

The kubernetes autodiscovery with its kubebuilder structure uses
[ComponentConfig][1] for this, which is deprecated and creates options to
be consumed by the manager, which this project does not use. To keep the
whole experience similar, the [kubernetes yaml package][2] is used to
provide the same parsing behaviour. It converts the file to json first,
which means structs do not need yaml tags in addition to json tags.

To keep code duplication minimal, the cloud autodiscovery reuses parts
of the kubernetes autodiscovery for the config. The ScanConfigs are part
of configv1 and the whole templating code is imported to provide the
same convenience of templated scan definitions. Some functions that are
privately defined in the controllers are copied over, in the future
these might get relocated to a shared util package.

The auto-discovery-cloud-aws-config.yaml file contains an example config
that can be used for local development. Other than for the kubernetes
autodiscovery just using this file is not enough though, because the
settings to connect to the SQS queue on AWS are individual. Those can
also be specified through environment variables, which override the
values specified in the config file.

[1]: https://kubebuilder.io/component-config-tutorial/tutorial
[2]: https://pkg.go.dev/sigs.k8s.io/yaml

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>