#1894 Add Helm Chart for Cloud-AWS AutoDiscovery
Add the necessary structure for making this a proper auto discovery
service: Makefile, Dockerfile, Helm Chart. The files are adapted from
the Kubernetes AutoDiscovery.

Since the Cloud-AWS AutoDiscovery does not need to monitor resources in
kubernetes and only creates scans, it does not need a ClusterRole, a
Role and RoleBinding for the namespace it creates the scans in is
enough. The current set of permissions is probably too large, this can
be restricted further later.

For connecting to AWS there are no good defaults, so other than the
Kubernetes AutoDiscovery, there cannot be a completely ready dev-config.
To ease that, the Cloud-AWS AutoDiscovery allows specifying AWS
connection options in multiple ways, either through the config or as
environment variables. When the helm-deploy Makefile target is used,
values for SQS_QUEUE_URL and AWS_REGION get picked up from the shell's
environment and set as Helm values. AWS credentials need to be provided
as a Kubernetes secret, which can be configured through values under
awsAuthentication. Credential values are optional, the service can also
connect to AWS using only part of the credentials or through IAM roles
if it is running on AWS itself.

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>