#1894 Add vscode profiles for debugging
This simplifies debugging the AutoDiscovery service when running it
independenly. Unfortunately it seems to be impossible to get
setup-envtest to install the binaries to a predictable path, the only
argument it accepts is --bin-dir, but it still appends /k8s/<os things>.
Moving the binaries manually is a bit of a pain. Vscode cannot set env
variables from script outputs though, so just running setup-envtest also
does not work.

To solve this, a preLaunchTask is used, which invokes a Makefile target
to create an environment file which contains the path setup-envtest
comes up with, which can then be read by vscode. Kind of unnecessary,
but works.

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>