#1894 Stop scanning very short lived containers
Sometimes when containers on ECS exit very quickly, they never enter the
RUNNING status and instead go directly from PENDING to STOPPED.
Previously the AutoDiscovery would track all the containers that were
PENDING and create individual scans for those that transitioned to
STOPPED. This was a bit of a hack though, because then those scans would
never get cleaned up and it would require additional configuration.

This instead removes the detection of short lived containers for the
MVP, it might get added back at a later point.

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>