#1894 Expand and reorder docs
The docs did not make clear enough that the AWS Cloud AutoDiscovery is
in a prerelease state and while it monitors AWS/ECS, it runs on
Kubernetes, in or out of AWS. This is added using the [GitHub
NOTE/WARNING syntax][1], since that is the main place where this info is
displayed, it is not used for the docs on the website. On DockerHub and
ArtifactHub it should be ok.

The order/structure of the README was not the best, it started with the
(default and autogenerated) helm install command, which would fail
anyway because the whole AWS setup is missing. The AWS setup was only
described after that. This reverses the order and moves all that setup
into a Prerequisites section, and also adds another helm command to the
Deployment section, that shows how to actually send the necessary values
to the deployment. Unfortunately it is not possible to get rid of the
default deployment command without altering all the templates.

This also clarifies some details, if all values are necessary and what
the AWS cost calculations are based on.

[1]: https://github.com/orgs/community/discussions/16925

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>