#1894 Hide the option to configure the namespace
Technically the namespace the scans get created in can be configured by
setting config.kubernetes.namespace, but outside of development it is of
low practical value, because the Helm chart only defines a Role and a
RoleBinding for the namespace the AutoDiscovery gets installed in. To
allow it to create scans in a different namespace, more
Roles/RoleBindings or a ClusterRole and a ClusterRoleBinding are needed.
These could be created manually but it is not the nicest way to do this,
so officially configuring the namespace is not supported for now.

To achieve this, the namespace is removed from values.yaml, so that it
is hidden in the documentation, but it can still be set from the config
passed with --config, which is needed for local development.

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>