#1894 Change golang structure
Standard or good golang project layout seems to be quite controversial,
there is the [Standard Go Project Layout][1], which comes with an
[extensive disussion][2] if this is encouraged or not. There are also
[blogposts][3] by go team members and the official [Organizing a Go
module][4] docs.

The takeaway I got was that a cmd dir is good for the executables and a
pkg dir is fine if it makes the repository structure better, and with
all the helm and docker files this is certainly the case.

[1]: https://github.com/golang-standards/project-layout
[2]: https://github.com/golang-standards/project-layout/issues/117
[3]: https://eli.thegreenplace.net/2019/simple-go-project-layout-with-modules/
[4]: https://go.dev/doc/modules/layout

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>