#1894 Improve docker reference handling
Move docker reference normalization to the kubernetes side, it fits
better there. Several places of the code need different parts of the
docker reference of the running containers. This simplifies the access
to those properties and uses them in a more consistent way. The scans
now only use the actual image name for their name, so that there is
space for information other than docker-io-. The hashmap of running
containers now uses a normalized reference string instead of the
ImageInfo objects, which makes it less error prone with the included
reference information.

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>