#1894 Remove unused functions and k8s permissions
The initial set of kubernetes permissions for the role in the helm
template was adapted from the kubernetes AutoDiscovery, which is based
on kubebuilder. The AWS Cloud AutoDiscovery manually connects to
kubernetes and uses a lot fewer API permissions than the kubernetes
AutoDiscovery. If future features need more permissions they can be
added back again.

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>