#2061 Read name and version from scan annotations
Instead of relying on the brittle docker image reference parsing, read
the project name and version for sending the SBOM to Dependency-Track
from annotations, simmilar to how other hooks (especially DefectDojo)
already do this. If the annotations are missing fall back to the regex.

The annotations use _project_ instead of _product_ because that is the
terminology Dependency-Track uses.

Signed-off-by: Lukas Fischer <lukas.fischer@iteratec.com>