Add a wait before refetching DefectDojo results.
It can happen that the results refetched from the DefectDojo API are
incorrect.
This only happens if the deduplication feature is enabled in DefectDojo.
The cause for this is that after importing a finding is deduplicated
in the background and depending on how performant your DefectDojo is,
this might take a couple of seconds to finish.

If the results are refetched before the deduplication finishes and the
status of all findings has been updated, the results we get back from
the API will not be correct.

The attributes which might be incorrect include but are not limited to
falsePositive, duplicate, riskAccepted and outOfScope.

Allowing the user to configure a sleep/wait in seconds helps to fix this
in flexible way. By allowing the user adjust this depending on the load
and performance on their DefectDojo installation.

Signed-off-by: Maximilian Dorner <maximilian.dorner@clever-soft.com>