Map the zap-automation-scan to the ZAP_SCAN type
Scans done with the ZAP automation framework were not mapped to any scan
type. That resulted in the usage of the Generic findings import, which
limits the features provided by DefectDojo (e.g. missing CWE links).

The results of the different ZAP scan methods all provide the same
results in the same format. This allows us to use the same scan type for
all of them.

Signed-off-by: Maximilian Dorner <maximilian.dorner@clever-soft.com>