#2057 Implements awaiting before refetching imported DefectDojo finding
The problem is that DefectDojo does the deduplication async. So, if we
import a finding and imediately search for it, it will be found, although
it will be deduplicated later on.

Since DefectDojo does not provide an API to query if dedup is done, we
add a wait as workaround. This is merely a sleep of the current thread
which is configurable via an env var:

```
DEFECTDOJO_REFETCH_WAIT_SECONDS=1
```

This config option is deprecated because we will remove it imediately
without substitue, if Defectdojo provides the necessary API.

Co-authored-by: Maximilian Dorner <maximilian.dorner@clever-soft.com>
Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>