#2272 Do Not Return Null
Null should be avoid whenever possible because it leads to endless
drama with NullPointerExceptions at runtime.

In this particular case it is formatted as a string which will
(maybe) convert to the string "null". Since this is an implementation
detail of the Java SDK we can rely on that.

Introduce the explicit string "n/a"" to indicate that there is no
starting time.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>