#2272 Fix NPE on Null Parameters From Scan Spec
This value may be null, but the join method requirres non-null.

The fix is to check for null and defult to an empty list. This
object creation should not be a performace problem. If to many
descritions wil be generated this can be circumvented by default
to a hardcoded string literal for empty/null parameters.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>