Migrate auto-discovery to newest kubebuilder version
This turned out to be a bit tricky because of the deprecation/removal of the controller runtime config handling.
I've moved the config handling to use viper now which seems to be working fine.
Had to move some stuff around for it, like he metav1.Duration to the default duration.
This actually made the code cleaner so no complaints from me.

Signed-off-by: Jannik Hollenbach <jannik.hollenbach@iteratec.com>