Handle Erros Parsing Kubernetes Config
Since the error handling of Kubernetes lib and the used YAML parser lib
is quite shitty the resulting error message is not very precise.
The problem is that e.g. for empty file they just throw an NPE on a
underlying Map<>. For bad format it just throw random errors such as
class cast excpetions.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>