Improve cascading scans docs by rewording and including a overview diagram
Put the diagram onto the website static directory and used a absolute link to it to ensure that it works whereever this is displayed.
It might be opened on either github, our website, artifacthub or dockerhub so the image can't be included using a relative link as we can only get the readme set there.

Signed-off-by: Jannik Hollenbach <jannik.hollenbach@iteratec.com>